move //cli/ops.rs to //cli/ops/mod.rs
factor out 'files' ops

factor out 'net' ops

factor out 'workers' ops

factor out 'compiler' ops

factor out 'os' ops

remove unneeded hyper imports

factor out 'process' ops

factor out 'repl' ops

factor out 'permissions' ops

factor out 'fs' ops

fix rebase

factor out 'fetch' and 'timers' ops

factor out rest of ops

remove unneeded pub modifiers
